+++ src/Exceptions/DynamicMutatorInvalidOperatorException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Mindtwo\DynamicMutators\Exceptions;
+
+class DynamicMutatorInvalidOperatorException extends DynamicMutatorException
+{
+}