+++ src/Exceptions/DynamicMutatorNotCallableException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Mindtwo\DynamicMutators\Exceptions;
+
+class DynamicMutatorNotCallableException extends DynamicMutatorException
+{
+}