+++ src/Exceptions/InvalidParameterException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Mindtwo\DynamicMutators\Exceptions;
+
+class InvalidParameterException extends DynamicMutatorException
+{
+}