+++ src/Exceptions/MutatorNotCallableException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace mindtwo\LaravelDynamicModelMutators\Exceptions;
-
-use Exception;
-
-class MutatorNotCallableException extends Exception
-{
-}