+++ src/Facades/Handler.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Mindtwo\DynamicMutators\Facades;
+
+use Illuminate\Support\Facades\Facade;
+use Mindtwo\DynamicMutators\Traits\HasMakeMapper;
+
+class Handler extends Facade
+{
+    use HasMakeMapper;
+}