+++ src/Model.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Mindtwo\DynamicMutators;
+
+use Mindtwo\DynamicMutators\Traits\HasDynamicMutators;
+
+class Model extends \Illuminate\Database\Eloquent\Model
+{
+    use HasDynamicMutators;
+}