+++ tests/TestCase.php
@@ -1,9 +1,7 @@
 <?php
 
-namespace mindtwo\LaravelDynamicModelMutators\Tests;
+namespace Mindtwo\DynamicMutators\Tests;
 
-use PHPUnit\Framework\TestCase as BaseTestCase;
-
-abstract class TestCase extends BaseTestCase
+abstract class TestCase extends \Orchestra\Testbench\TestCase
 {
 }