+++ CHANGELOG.md
@@ -2,10 +2,22 @@
 
 All notable changes to `laravel-dynamic-model-mutators` will be documented in this file
 
+## 2.0.0
+
+- Complete refactoring
+- Integrated handler object
+- Use any callable as mutator function
+- Stack mutator functions
+- Use eloquent's append functionality 
+- Integrated mutation handler facade
+- Make mapper to initialize handler with arguments in custom order
+
+
 ## 1.1.1 - 2018-12-03
 
 - Fixed error in hasDynamicMutator() method
 - Test namespace refactored
+
 
 ## 1.1.0 - 2018-12-03
 
@@ -15,11 +27,13 @@ All notable changes to `laravel-dynamic-model-mutators` will be documented in th
 - Added mutator operator not defined exception
 - Extended documentation  
 
+
 ## 1.0.1 - 2018-11-29
 
 - Unit tests for laravel's fillable feature
 - Updated readme
 - Comments fixed
+
 
 ## 1.0.0 - 2018-11-20